Move Link and Node to separate files
