+++ backend/drawables/Node.cs
@@ -1,0 +1,9 @@
+namespace backend.drawables
+{
+   public class Node 
+    {
+        public string server_id { get; set; }
+        public string server_name { get; set; }
+        public bool ntv_error { get; set; }
+    }
+}
